Removing files since MemoryExtensions is in CoreLib now.
